redesign verify member sheet
