+++ khelo/lib/ui/flow/team/add_team_member/add_team_member_screen.dart
@@ -12,7 +12,7 @@ import 'package:khelo/components/user_detail_cell.dart';
 import 'package:khelo/domain/extensions/context_extensions.dart';
 import 'package:khelo/ui/flow/matches/add_match/select_squad/components/user_detail_sheet.dart';
 import 'package:khelo/ui/flow/team/add_team_member/add_team_member_view_model.dart';
-import 'package:khelo/ui/flow/team/add_team_member/components/verify_add_team_member_dialog.dart';
+import 'package:khelo/ui/flow/team/add_team_member/components/verify_team_member_sheet.dart';
 import 'package:style/animations/on_tap_scale.dart';
 import 'package:style/button/action_button.dart';
 import 'package:style/extensions/context_extensions.dart';
@@ -116,7 +116,7 @@ class _AddTeamMemberScreenState extends ConsumerState<AddTeamMemberScreen> {
                             !state.selectedUsers.contains(user)
                         ? () async {
                             if (user.phone != null) {
-                              final res = await VerifyAddTeamMemberDialog.show(
+                              final res = await VerifyTeamMemberSheet.show(
                                   context,
                                   phoneNumber: user.phone!);
                               if (res != null && res && context.mounted) {
